Move Sysctl class to ::CFA namespace
